% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_olap.R
\name{tag_olap}
\alias{tag_olap}
\title{Calculate Tag Overlap Statistic}
\usage{
tag_olap(
  ccamlr_data,
  taxon = NA,
  seasons = NULL,
  vessels = NULL,
  areas = "ALL",
  plot = F
)
}
\arguments{
\item{ccamlr_data}{A mandatory parameter representing the official CCAMLR data extract. The function relies on the presence of five specific data frames in the extract: C2, C2_CATCH, OBS_HAUL_BIOLOGY, OBS_HAUL_TAG_RELEASE, OBS_HAUL_TAG_RECAPTURE.}

\item{taxon}{A mandatory parameter specifying the taxon code for which the overlap statistic is calculated. Options are 'TOP' or 'TOA'.}

\item{seasons}{An optional parameter with a default value of NULL. Specify a list of seasons for which you want to calculate the overlap statistic, e.g., c(2020, 2021, 2022). The default value of NULL calculates the overlap statistics for all seasons present in the data.}

\item{vessels}{An optional parameter with a default value of NULL. Specify a list of vessels for which you want to calculate the overlap statistic, e.g., c('Tronio', 'Janas'). The default value of NULL calculates the overlap statistics for all vessels present in the data.}

\item{areas}{An optional parameter with a default value of 'ALL'. Specify a list of asd_codes, ssru_codes, or rb_codes for which you want to calculate the overlap statistic, e.g., c('881', '881H', '881K', '486_2', 'combined'). Use 'combined' to calculate a single combined statistic for all areas. Use 'CCEP' to calculate the overlap statistic solely for areas where a CM requires a minimum overlap. The default value of 'ALL' calculates the overlap statistics for all possible areas present in the data.}

\item{plot}{An optional parameter with a default value of FALSE. When set to TRUE, the tag overlap statistic function will generate a plot for each calculated overlap statistic.}
}
\description{
This function computes the tag overlap statistic and tagging rate per vessel based on the provided CCAMLR data extract.
}
\examples{
tag_olap(ccamlr_data = ccamlr_data, taxon = 'TOP', seasons = 2017, areas = '881')
tag_olap(ccamlr_data = ccamlr_data, taxon = 'TOA', seasons = c(2016, 2017), areas = 'CCEP')
tag_olap(ccamlr_data = ccamlr_data, taxon = 'TOA', vessels = c('vessel_2'), seasons = c(2016, 2017), areas = c("881"), plot = TRUE)

}
