% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_olap.R
\name{tag_olap}
\alias{tag_olap}
\title{Calculate Tag Overlap Statistic}
\usage{
tag_olap(
  ccamlr_data,
  taxon = NA,
  seasons = NULL,
  vessels = NULL,
  areas = "ALL",
  custom_areas = NULL,
  ccep_season = NULL,
  plot = F
)
}
\arguments{
\item{ccamlr_data}{A required data extract provided by CCAMLR. The extract must contain five specific data frames: `C2`, `C2_CATCH`, `OBS_HAUL_BIOLOGY`, `OBS_HAUL_TAG_RELEASE`, and `OBS_HAUL_TAG_RECAPTURE`. These data frames are essential for calculating the overlap statistic.}

\item{taxon}{A required parameter indicating the taxon code for which the overlap statistic will be calculated. Accepted values are 'TOP' or 'TOA'.}

\item{seasons}{Optional. A list of seasons for which to calculate the overlap statistic (e.g., `c(2020, 2021, 2022)`). If `NULL` (default), the function calculates the overlap for all seasons present in the data.}

\item{vessels}{Optional. A list of vessels for which to calculate the overlap statistic (e.g., `c('Tronio', 'Janas')`). If `NULL` (default), the function calculates the overlap for all vessels in the dataset.}

\item{areas}{Optional. A list of area codes (e.g., asd_codes, ssru_codes, or rb_codes such as `c('881', '881H', '881K', '486_2')`). Use `'combined'` to compute a single combined statistic for all areas or `'CCEP'` for areas requiring a minimum overlap according to Conservation Measures (CM). The default value is `'ALL'`, which calculates the overlap for all available areas. To define custom areas, set `areas = 'custom'` and provide the column name via the `custom_area` parameter.}

\item{custom_areas}{Optional. Required when `areas = 'custom'`. This parameter must specify the name of the column in the `C2` table that contains the custom areas for which you want to calculate the overlap statistic.}

\item{plot}{Optional. A logical flag (default is `FALSE`). When `TRUE`, the function generates a plot for each calculated overlap statistic.}
}
\description{
This function computes the tag overlap statistic and tagging rate per vessel based on the provided CCAMLR data extract.
}
\examples{
tag_olap(ccamlr_data = ccamlr_data, taxon = 'TOP', seasons = 2017, areas = '881')
tag_olap(ccamlr_data = ccamlr_data, taxon = 'TOA', seasons = c(2016, 2017), areas = 'CCEP')
tag_olap(ccamlr_data = ccamlr_data, taxon = 'TOA', vessels = c('vessel_2'), seasons = c(2016, 2017), areas = c("881"), plot = TRUE)

}
